LU-1996 lustre: Flexible changelog format.
Added jobid fields to Changelog records (and extended records). The
CLF_JOBID flags allows to check if the field is present or not (old
format) when reading an entry. Jobids are expressed as 32 chars long,
zero-terminated strings. Updated test_205 in sanity.sh.

+++ lustre/tests/sanity.sh
@@ -11245,6 +11245,11 @@ test_205() { # Job stats
 	cmd="mv -f $DIR/$tfile $DIR/jobstats_test_rename"
 	verify_jobstats "$cmd" "mdt"
 
+	# Ensure that jobid are present in changelog
+	$LFS changelog $MDT0 | tail -9
+	jobids=$($LFS changelog $MDT0 | tail -9 | grep -c "j=")
+	[ $jobids -eq 9 ] || error "Wrong changelog jobid count $jobids != 9"
+
 	# cleanup
 	rm -f $DIR/jobstats_test_rename
